Move Fingerscans to the sbt-plugin module
