+++ sbt/plugin/src/main/java/ch/epfl/scala/debugadapter/sbtplugin/SubclassFingerscan.java
@@ -5,7 +5,7 @@
  * Licensed under Apache License 2.0 (see LICENSE)
  */
 
-package ch.epfl.scala.debugadapter.sbtplugin.internal;
+package ch.epfl.scala.debugadapter.sbtplugin;
 
 import sbt.testing.SubclassFingerprint;
